src/research/jwst_pipeline/requirements.txt
astroquery
astropy
numpy
psycopg2-binary
python-dotenv
